Moving the player

+++ src/game/fruit.rs
@@ -11,8 +11,8 @@ impl Fruit {
     pub fn new(rng: &mut rand::prelude::ThreadRng, width: u32, height: u32) -> Fruit {
         Fruit {
             center: Position {
-                x: rng.gen_range(0..width as i32),
-                y: rng.gen_range(0..height as i32),
+                x: rng.gen_range(0..width) as f32,
+                y: rng.gen_range(0..height) as f32,
             },
             radius: settings::FRUIT_RADIUS,
         }
